refactor: isolate catalog renderer feature
